// src/api/typesGenerated.ts
export type Visitor = {
  ip: string,
  loginUsers: LoginUser[]
  workspaceUsers: WorkspaceUser[]
}
export type LoginUser = {
  user_id: string,
  fullname: string,
  login_time: string
}
export type WorkspaceUser = {
  ws_name: string,
  owner_name: string,
  time: string,
}